% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignSomy.R
\name{assign_gainloss}
\alias{assign_gainloss}
\title{Assign CNV state}
\usage{
assign_gainloss(seq_data, cluster, uq = 0.9, lq = 0.1)
}
\arguments{
\item{seq_data}{Sequential data - Counts per bin}

\item{cluster}{Vector showing segment identity}

\item{uq}{Upper quantile to trim to calculate the cluster means}

\item{lq}{Lower quantile to trim to calculate the cluster means}
}
\value{
Copy number states for the different segments
}
\description{
Assign CNV state
}
