% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotype_measures.R
\name{cnv_burden_per_cell}
\alias{cnv_burden_per_cell}
\title{Calculates CNV burden per cell}
\usage{
cnv_burden_per_cell(res_table, base_state = 1)
}
\arguments{
\item{res_table}{Result data frame generated by the epiAneufinder main function}

\item{base_state}{Numeric encoding of base state (default 1)}
}
\value{
Aneuploidy (=CNV burden) per cell as a named vector
}
\description{
The CNV burden (=aneuploidy score) per cell is calculated as 
the mean deviation from the base state across all bins
}
