% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{countInsertions}
\alias{countInsertions}
\title{Creates the count matrix using either a fragment file or one bam file per cell}
\usage{
countInsertions(reads, windows, by = NULL, minFrags = NULL, mapqFilter = NULL)
}
\arguments{
\item{reads}{BamFileList, Grange object or dgCMatrix object dependent 
on the original input file type (see main wrapper function)}

\item{windows}{binned genome as generated by the makeWindows function}

\item{by}{barcode information}

\item{minFrags}{Minimum number of fragments a barcode must contain to be counted as a cell}

\item{mapqFilter}{Filter bam files after a certain mapq value}
}
\description{
Creates the count matrix using either a fragment file or one bam file per cell
}
