% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceStatistic.R
\name{dist_ad}
\alias{dist_ad}
\title{Function to calculate the AD statistic between two distributions}
\usage{
dist_ad(x, y, test = "AD")
}
\arguments{
\item{x}{Vector with reads counts left of the breakpoint}

\item{y}{Vector with reads counts right of the breakpoint}

\item{test}{Test statistics (either AD, KS or Bhattacharya)}
}
\description{
Function to calculate the AD statistic between two distributions
}
