% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCountMatrix.R
\name{generateCountMatrix}
\alias{generateCountMatrix}
\title{Generate count matrix}
\usage{
generateCountMatrix(reads, windows, by = NULL, minFrags = NULL)
}
\arguments{
\item{reads}{bamfile or fragments file of sequencing information}

\item{windows}{Binned genome}
}
\description{
Generate count matrix
}
