% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBreakpoints.R
\name{getbp}
\alias{getbp}
\title{Identify breakpoints}
\usage{
getbp(seq_data, minsize = 1, k = 3, test = "AD", minsizeCNV = 5)
}
\arguments{
\item{seq_data}{Sequential data - Counts per bin for a chromosome}

\item{minsize}{Integer. Resolution at the level of ins. Default: 1. Setting it to higher numbers runs the algorithm faster at the cost of resolution}

\item{k}{Integer. Find 2^k segments per chromosome}

\item{test}{Test statistics}

\item{minsizeCNV}{Integer. Number of consecutive bins to constitute a possible CNV}
}
\description{
Identify breakpoints
}
