% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotype_measures.R
\name{karyotype_measures}
\alias{karyotype_measures}
\title{Calculates karyotype heterogeneity and aneuploidy from the result CNVs}
\usage{
karyotype_measures(
  res_table,
  base_state = 1,
  plot_path = NULL,
  plot_width = 5,
  plot_height = 4
)
}
\arguments{
\item{res_table}{Result data frame generated by the epiAneufinder main function}

\item{base_state}{Numeric encoding of base state (default 1)}

\item{plot_path}{Path to save a scatter plot of chromosome aneuploidy vs heterogeneity, 
only done if plot_path is not NULL (default NULL)}

\item{plot_width}{Width of the plot}

\item{plot_height}{Height of the plot}
}
\value{
Aneuploidy and heterogeneity genome-wide and per chromosome
}
\description{
The aneuploidy score is defined as the mean deviation from the base state across all cells and bins.
The heterogeneity score is defined as in the package Aneufinder (see Baker et al, Genome Biology 2016).
}
