% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWindows.R
\name{makeWindows}
\alias{makeWindows}
\title{Create windows (=bins) over the complete genome}
\usage{
makeWindows(genome, blacklist, windowSize, exclude = NULL)
}
\arguments{
\item{genome}{String containing name of BS.genome object. Necessary for GC correction. Default: "BSgenome.Hsapiens.UCSC.hg38"}

\item{blacklist}{Granges object with information about blacklisted regions in the genome.}

\item{windowSize}{Size in basepairs for a bin}

\item{exclude}{Chromosomes to exclude}
}
\description{
Create windows (=bins) over the complete genome
}
