% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_cnv_results.R
\name{plot_karyo_annotated}
\alias{plot_karyo_annotated}
\title{Recreate the karyogram plot - optional with an annotation side bar}
\usage{
plot_karyo_annotated(res_table, plot_path, annot_dt = NULL, title_karyo = "")
}
\arguments{
\item{res_table}{Result data frame generated by the epiAneufinder main function}

\item{plot_path}{Path under which the karyogram is saved}

\item{annot_dt}{Optional an annotation data frame with the columns cell (=cell barcode) 
and annot (= categories or numerical values for the annotation side bar). Set value to NULL
if no side bar should be plotted.}

\item{title_karyo}{Plot title for the karyogram}
}
\description{
Recreate the karyogram plot - optional with an annotation side bar
}
