% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_karyo_gainloss}
\alias{plot_karyo_gainloss}
\title{Plot karyogram}
\usage{
plot_karyo_gainloss(
  somies_ad,
  outdir,
  peaks,
  uq = NULL,
  lq = NULL,
  title_karyo = NULL
)
}
\arguments{
\item{somies_ad}{A list containing the somy per bin of each cell}

\item{outdir}{Directory where the output karyogram is to be saved}

\item{peaks}{Dataframe containing bin information}

\item{uq}{Upper quantile. Default: 0.1}

\item{lq}{Lower quantile. Default: 0.9}

\item{title_karyo}{String. Title of the output karyogram}
}
\description{
Plot karyogram
}
