% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_cnv_results.R
\name{plot_single_cell_profile}
\alias{plot_single_cell_profile}
\title{Split cells into different subclones based on hierarchical clustering of the CNV profiles}
\usage{
plot_single_cell_profile(
  outdir,
  threshold_blacklist_bins = 0.85,
  selected_cells = 10,
  plot_path = "individual_tracks.pdf",
  plot_width = 25,
  plot_height = 15
)
}
\arguments{
\item{outdir}{Directory created by the epiAneufinder main function with required
files count_summary.rds, counts_gc_corrected.rds and and cnv_calls.rds}

\item{threshold_blacklist_bins}{Blacklist a bin if more than the given ratio of cells have zero reads in the bin. Default: 0.85}

\item{selected_cells}{Either a character vector with barcodes of cells to visualize or an integer number to visualize
a random subset of this size}
}
\description{
Split cells into different subclones based on hierarchical clustering of the CNV profiles
}
