% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_correction_helper.R
\name{quadratic_gc_correction}
\alias{quadratic_gc_correction}
\title{Help function for GC correction per cell, in case that instead of the loess function, 
a standard implemetation is used with a quadratic fit for each polynome}
\usage{
quadratic_gc_correction(
  counts,
  gc_content,
  intervals_per_bin,
  mean_gc_interval
)
}
\arguments{
\item{counts}{Vector with counts of one cell across all bins}

\item{gc_content}{Vector with gc content per bin}

\item{intervals_per_bin}{Vector assigning bin to GC interval}

\item{mean_gc_interval}{Mean GC content of each interval}
}
\description{
Help function for GC correction per cell, in case that instead of the loess function, 
a standard implemetation is used with a quadratic fit for each polynome
}
