% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readCountMatrix}
\alias{readCountMatrix}
\title{Read directly a count matrix saved in mtx format}
\usage{
readCountMatrix(dir_path)
}
\arguments{
\item{dirpath}{Directory path that needs to contains three files in that directory:
matrix.mtx, barcodes.tsv and peaks.bed}
}
\description{
Read directly a count matrix saved in mtx format
}
