% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceStatistic.R
\name{seq_dist_ad}
\alias{seq_dist_ad}
\title{Function to calculate the breakpoints with AD stat given a series of data points}
\usage{
seq_dist_ad(seq_data, minsize = 3, test = "AD")
}
\arguments{
\item{seq_data}{Normalized read counts per window}

\item{minsize}{Integer. Resolution at the level of ins. Default: 1. Setting it to higher numbers runs the algorithm faster at the cost of resolution}

\item{test}{Test statistics (either AD, KS or Bhattacharya)}
}
\description{
Function to calculate the breakpoints with AD stat given a series of data points
}
