% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_cnv_results.R
\name{split_subclones}
\alias{split_subclones}
\title{Split cells into different subclones based on hierarchical clustering of the CNV profiles}
\usage{
split_subclones(
  res_table,
  tree_depth,
  plot_tree = TRUE,
  plot_path = "dendogram_clustering.pdf",
  plot_width = 4,
  plot_height = 3
)
}
\arguments{
\item{res_table}{Result data frame generated by the epiAneufinder main function}

\item{tree_depth}{Depth at which the dendogram should be cut}

\item{plot_tree}{Boolean value whether the cutted dendogram should be visualized}

\item{plot_path}{Path under which the dendogram is saved as a png (only for plot_tree=TRUE)}

\item{plot_width}{Width of the png (only for plot_tree=TRUE)}

\item{plot_height}{Height of the png (only for plot_tree=TRUE)}
}
\description{
Split cells into different subclones based on hierarchical clustering of the CNV profiles
}
