% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignSomy.R
\name{threshold_dist_values}
\alias{threshold_dist_values}
\title{Prunes breakpoints}
\usage{
threshold_dist_values(result.dt)
}
\arguments{
\item{results}{result.dt A data table combining the breakpoints and distances}
}
\value{
A data table after pruning the noisy breakpoints
}
\description{
Prunes breakpoints
}
