% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{countInsertions}
\alias{countInsertions}
\title{Creates the count matrix using either a fragment file or one bam file per cell}
\usage{
countInsertions(reads, windows, by = NULL, minFrags = NULL)
}
\arguments{
\item{reads}{bamfile or fragments file}

\item{windows}{binned genome as generated by the makeWindows function}

\item{by}{barcode information}

\item{minFrags}{Minimum number of fragments a barcode must contain to be counted as a cell}
}
\description{
Creates the count matrix using either a fragment file or one bam file per cell
}
