% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiAneufinder_function.R
\name{epiAneufinder}
\alias{epiAneufinder}
\title{Wrapper function for the \code{\link{epiAneufinder}} package}
\usage{
epiAneufinder(
  input,
  outdir,
  blacklist,
  windowSize,
  genome = "BSgenome.Hsapiens.UCSC.hg38",
  test = "AD",
  reuse.existing = FALSE,
  exclude = NULL,
  uq = 0.9,
  lq = 0.1,
  title_karyo = NULL,
  minFrags = 20000,
  mapqFilter = 10,
  threshold_cells_nbins = 0.05,
  selected_cells = NULL,
  gc_correction = "quadratic",
  threshold_blacklist_bins = 0.85,
  ncores = 4,
  minsize = 1,
  k = 4,
  minsizeCNV = 0,
  plotKaryo = TRUE
)
}
\arguments{
\item{input}{Folder with bam files, a fragments.tsv/bed file or a folder with
a count matrix (required files: matrix.mtx(.gz), barcodes.tsv(.gz) and peaks.bed(.gz))}

\item{outdir}{Path to output directory}

\item{blacklist}{Bed file with blacklisted regions}

\item{windowSize}{Size of the window (Reccomended for sparse data - 1e6)}

\item{genome}{String containing name of BS.genome object. Necessary for GC correction. Default: "BSgenome.Hsapiens.UCSC.hg38"}

\item{test}{One of AD or KS. Anderson-Darling or Kolmogorov-Smirnov. Default: "AD"}

\item{reuse.existing}{Logical. False removes all the files in the outdir and recomputes everything.}

\item{exclude}{String of chromosomes to exclude. Example: c('chrX','chrY','chrM')}

\item{uq}{Upper quantile. Default: 0.1}

\item{lq}{Lower quantile. Default: 0.9}

\item{title_karyo}{String. Title of the output karyogram}

\item{minFrags}{Integer. Minimum number of reads for a cell to pass. Only required for fragments.tsv file. Default: 20000}

\item{mapqFilter}{Filter bam files after a certain mapq value}

\item{threshold_cells_nbins}{Keep only cells that have more than a certain percentage of non-zero bins}

\item{selected_cells}{Additional option for filtering the input, either NULL or a file with barcodes of cells to keep (one barcode per line, no header)}

\item{gc_correction}{Type of GC correction, currently implemented options are "loess", "bulk_loess" and "quadratic".}

\item{threshold_blacklist_bins}{Blacklist a bin if more than the given ratio of cells have zero reads in the bin. Default: 0.85}

\item{ncores}{Number of cores for parallelization. Default: 4}

\item{minsize}{Integer. Resolution at the level of ins. Default: 1. Setting it to higher numbers runs the algorithm faster at the cost of resolution}

\item{k}{Integer. Find 2^k segments per chromosome}

\item{minsizeCNV}{Integer. Number of consecutive bins to constitute a possible CNV}

\item{plotKaryo}{Boolean variable. Whether the final karyogram is plotted at the end}
}
\value{
\code{NULL}
}
\description{
Wrapper function for the \code{\link{epiAneufinder}} package
}
\author{
Akshaya Ramakrishnan
}
