% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCountMatrix.R
\name{generateCountMatrix}
\alias{generateCountMatrix}
\title{Generate count matrix}
\usage{
generateCountMatrix(
  reads,
  windows,
  by = NULL,
  minFrags = NULL,
  mapqFilter = NULL
)
}
\arguments{
\item{reads}{BamFileList, Grange object or dgCMatrix object dependent 
on the original input file type (see main wrapper function)}

\item{windows}{Binned genome}

\item{by}{barcode information}

\item{minFrags}{Minimum number of fragments a barcode must contain to be counted as a cell}

\item{mapqFilter}{Filter bam files after a certain mapq value}
}
\description{
Generate count matrix
}
