% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotype_measures.R
\name{karyotype_measures}
\alias{karyotype_measures}
\title{Calculates karyotype heterogeneity and aneuploidy from the result CNVs}
\usage{
karyotype_measures(res_table, base_state = 1)
}
\arguments{
\item{res_table}{Result data frame generated by the epiAneufinder main function}

\item{base_state}{Numeric encoding of base state (usually 1)}
}
\value{
Aneuploidy and heterogeneity genome-wide and per chromosome
}
\description{
The aneuploidy score is defined as the mean deviation from the base state across all cells and bins.
The heterogeneity score is defined as in the package Aneufinder (see Baker et al, Genome Biology 2016).
}
