% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readCountMatrix}
\alias{readCountMatrix}
\title{Read directly a count matrix saved in mtx format}
\usage{
readCountMatrix(dir_path)
}
\arguments{
\item{dirpath}{Directory path that needs to contains three files in that directory:
matrix.mtx(.gz), barcodes.tsv(.gz) and peaks.bed(.gz)}
}
\description{
Read directly a count matrix saved in mtx format
}
